requirements.txt
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.173.0
python-telegram-bot==22.1
requests==2.32.4
python-dotenv==1.1.0
cohere==4.47 